// packages/core/src/types.ts
export type BuildInfo = {
    version?: string
    git_tag?: string
    git_branch?: string
    git_commit?: string
    build_time?: number
    environment?: string
    runtime?: string
    tags?: Record<string, string>
}